Trait chain_core::property::DeserializeFromSlice

pub trait DeserializeFromSlice {
    fn deserialize_from_slice(
        codec: &mut Codec<&[u8]>
    ) -> Result<Self, ReadError>;

    fn deserialize_validate_from_slice(
        codec: &mut Codec<&[u8]>
    ) -> Result<(), ReadError> { ... }
}

Define that an object can be read from a byte slice. This trait is implemented for all Deserialize implementors by default. The default implementation can be overridden if the user is sure they can benefit from slice-specific functions of Codec.

Required Methods

fn deserialize_from_slice(codec: &mut Codec<&[u8]>) -> Result<Self, ReadError>

Provided Methods

fn deserialize_validate_from_slice(
codec: &mut Codec<&[u8]>
) -> Result<(), ReadError>