Trait chain_core::property::Fragment

pub trait Fragment: Serialize {
    type Id: FragmentId;

    fn id(&self) -> Self::Id;
}

A fragment is some item contained in a block, such as a transaction, a delegation-related certificate, an update proposal, and so on. Fragments can be serialized (so that they can be concatenated to form a binary block( and have a unique ID (typically the hash of their serialization). TODO: add a Readable trait bound

Required Associated Types

type Id: FragmentId

Required Methods

fn id(&self) -> Self::Id

Return the message’s identifier.