Trait chain_core::property::FromStr

pub trait FromStr: Sized {
    type Error: Error + Send + Sync + 'static;

    fn from_str(s: &str) -> Result<Self, Self::Error>;
}

Defines the way to parse the object from a UTF-8 string.

This is like the standard FromStr trait, except that it imposes additional bounds on the error type to make it more usable for aggregation to higher level errors and passing between threads.

Required Associated Types

type Error: Error + Send + Sync + 'static

Required Methods

fn from_str(s: &str) -> Result<Self, Self::Error>